requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
fonttools==4.51.0
idna==3.7
importlib_resources==6.4.0
joblib==1.4.0
kiwisolver==1.4.5
matplotlib==3.7.2
numpy==1.22.0
opencv-python==4.8.0.74
packaging==24.0
pillow==10.3.0
pyparsing==3.0.9
python-dateutil==2.9.0.post0
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
-e git+ssh://git@github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588#egg=segment_anything
six==1.16.0
threadpoolctl==3.4.0
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.11.0
urllib3==2.2.1
zipp==3.18.1
